* topic/prebuild.sync.2015-05-20:
  Document tags [s-roff] and [s-roff-sync.1]
  Sync-to-go: doc (less info manual)
  Sync-to-go: man..
  Sync-to-go: tmac..
  Sync-to-go: src/ute-tfmtodit..
  Sync-to-go: src/ute-afmtodit..
  Sync-to-go: src/src-troff..
  Sync-to-go: src/src-roff..
  Sync-to-go: src/pre-tbl..
  Sync-to-go: src/pre-soelim..
  Sync-to-go: src/pre-preconv..
  Sync-to-go: src/pre-pic..
  Sync-to-go: src/pre-html..
  Sync-to-go: src/pre-grn..
  Sync-to-go: src/pre-eqn..
  Sync-to-go: src/dev-tty..
  Sync-to-go: src/dev-ps..
  Sync-to-go: src/dev-html..
  Sync-to-go: font/..
  Sync-to-go: contrib/pdfmark..
  Sync-to-go: contrib/hdtbl..
  s-gnu-diff.sh: easy compare codebase(s)
  Sync-to-go: update copyright for 2015

+++ src/troff/number.cpp
@@ -1,5 +1,5 @@
 /*@
- * Copyright (c) 2014 Steffen (Daode) Nurpmeso <sdaoden@users.sf.net>.
+ * Copyright (c) 2014 - 2015 Steffen (Daode) Nurpmeso <sdaoden@users.sf.net>.
  *
  * Copyright (C) 1989 - 1992, 2001, 2002, 2004
  *    Free Software Foundation, Inc.
@@ -359,7 +359,7 @@ static int parse_expr(units *v, int scale_indicator,
     case '*':
       if (v2 < 0) {
 	if (*v > 0) {
-	  if (*v > -(unsigned)INT_MIN / -(unsigned)v2)
+	  if ((unsigned)*v > -(unsigned)INT_MIN / -(unsigned)v2)
 	    overflow = 1;
 	}
 	else if (-(unsigned)*v > INT_MAX / -(unsigned)v2)
